% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prepareDataForReport}
\alias{prepareDataForReport}
\title{prepareDataForReport}
\usage{
prepareDataForReport(plinkBed, outFile = paste0(plinkBed,
  ".GenoTypingQC.html"))
}
\arguments{
\item{plinkBed}{1212}

\item{outFile}{21212}
}
\description{
prepareDataForReport
}

